Move tests files to be with source files
